Added GET to auth for getting user's password
This really needs to be handled differently. It gets the job done for now though, the password is sent as a response then sending the password is handled within the application in future itorations it would be better idea to handle this through the use of a one-time login token that is put in place of the password in the Users table then once the user logs in they are prompted to reset their password to a new one.